refactor: delete packages
