+++ app/src/main/kotlin/model/algo/findBridges.kt
@@ -1,4 +1,4 @@
-package model.algo.findBridges
+package model.algo
 
 import model.graph.Edge
 import model.graph.Graph
